+++ fractal_server/app/routes/api/v2/submit.py
@@ -220,6 +220,8 @@ async def apply_workflow(
     # Define user-side job directory
     if FRACTAL_RUNNER_BACKEND == "local":
         WORKFLOW_DIR_REMOTE = WORKFLOW_DIR_LOCAL
+    elif FRACTAL_RUNNER_BACKEND == "local_experimental":
+        WORKFLOW_DIR_REMOTE = WORKFLOW_DIR_LOCAL
     elif FRACTAL_RUNNER_BACKEND == "slurm":
         WORKFLOW_DIR_REMOTE = (
             Path(user.cache_dir) / f"{WORKFLOW_DIR_LOCAL.name}"