+++ fractal_server/app/routes/aux/_runner.py
@@ -5,15 +5,15 @@ from ....config import get_settings
 from ....syringe import Inject
 
 
-def _check_backend_is_slurm():
+def _is_shutdown_available():
     """
     Raises:
         HTTPException(status_code=HTTP_422_UNPROCESSABLE_ENTITY):
-            If FRACTAL_RUNNER_BACKEND is not 'slurm'
+            If FRACTAL_RUNNER_BACKEND is the thread-based 'local' backend.
     """
     settings = Inject(get_settings)
     backend = settings.FRACTAL_RUNNER_BACKEND
-    if backend != "slurm":
+    if backend not in ["slurm", "local_experimental"]:
         raise HTTPException(
             status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
             detail=(