+++ fractal_server/config.py
@@ -310,7 +310,9 @@ class Settings(BaseSettings):
             )
         return FRACTAL_RUNNER_WORKING_BASE_DIR_path
 
-    FRACTAL_RUNNER_BACKEND: Literal["local", "slurm"] = "local"
+    FRACTAL_RUNNER_BACKEND: Literal[
+        "local", "local_experimental", "slurm"
+    ] = "local"
     """
     Select which runner backend to use.
     """